Refactor persistence classes to be generic
Refactored `SQLitePersistence`, `FilePersistence`, and `InMemoryPersistence` to be generic, now named `SQLitePersistence<T>`, `FilePersistence<T>`, and `InMemoryPersistence<T>`, respectively. Updated `IPersistence` interface to be generic. Introduced `EntityFactory` for creating `IEntity` instances. Updated `TangoBotServiceProviderExp`, `StartUp`, and `HistoryDataStreamObserver` to handle new generic types. Enhanced `IEntity` and related classes for the new model. Updated unit tests and added utility methods for table handling and serialization.

+++ TangoBotTests/SQLitePersistenceTests.cs
@@ -11,7 +11,7 @@ namespace DatabaseLib.Tests
 {
     public class SQLitePersistenceTests : IDisposable
     {
-        private readonly IPersistence _persistence;
+        private readonly IPersistence<User> _persistence;
         private readonly string _dataDirectory;
         private readonly string _databaseFilePath;
 
@@ -24,11 +24,12 @@ namespace DatabaseLib.Tests
             //TangoBotServiceProvider.AddService<IPersistence>(provider => new SQLitePersistence(), typeof(SQLitePersistence).Name);
             //_persistence = TangoBotServiceProvider.GetSingletonService<IPersistence>(typeof(SQLitePersistence).Name) as SQLitePersistence ?? throw new System.Exception("Service not found");
 
-            _persistence = TangoBotServiceProviderExp.GetSingletonService<IPersistence>("DatabaseLib.SQLitePersistence") ?? throw new System.Exception("Service not found");
+            _persistence = TangoBotServiceProviderExp.GetSingletonService<IPersistence<User>>("DatabaseLib.SQLitePersistence") ?? throw new System.Exception("Service not found");
 
             try
             {
-                Directory.Delete(_databaseFilePath, true);
+                if(Directory.Exists(_databaseFilePath))
+                    Directory.Delete(_databaseFilePath, true);
             }
             catch (Exception)
             {
